refactor(sdk): keep operation partitions test-only
